backend/requirements.txt
# ===== requirements.txt =====

# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Pydantic
pydantic==2.5.3
pydantic-settings==2.1.0

# Browser automation
playwright==1.41.0

# Async support
aiofiles==23.2.1

# WebSocket
websockets==12.0

# Logging
colorlog==6.8.0

# Utils
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.20
asyncpg==0.27.1  # or latest compatible version

# Auth
email-validator==2.1.0.post1
python-jose[cryptography]==3.3.0

# System Stats
psutil==5.9.8
